+++ rayTracer/ObjectClasses/camera.cpp
@@ -9,6 +9,6 @@
 
 Camera::Camera()
 : Object(Vector3(0, 0, 0), Material()) {}
-float Camera::parameterize(const Vector3& origin, const Vector3& direction) const {
+double Camera::parameterize(const Vector3& origin, const Vector3& direction) const {
     return INFINITY;
 }
